Move AGFS queries to there on namespace module
So we can add LGFS namespace queries.
