+++ app/services/stats/management_information/queries/agfs/af2_high_value_query.rb
@@ -1,0 +1,36 @@
+# frozen_string_literal: true
+
+# https://docs.google.com/spreadsheets/d/1LgECNzbhOJ0MPS-2jbsXVyvtS2Q2TnjDQ4o42tyhU_g/edit#gid=1744254165
+# Where scheme = AGFS
+# And Where submission type = redetermination
+# And Where originally submitted = DATE specified for this lookup *
+# And Where disk evidence = no
+# And Where claim total > 20,000
+#
+
+require_relative '../base_query'
+
+module Stats
+  module ManagementInformation
+    module Agfs
+      class Af2HighValueQuery < BaseQuery
+        private
+
+        def query
+          <<~SQL
+            WITH journeys AS (
+              #{journeys_query}
+            )
+            SELECT count(*)
+            FROM journeys j
+            WHERE j.scheme = '#{@scheme}'
+            AND date_trunc('day', j.original_submission_date) = '#{@day}'
+            AND j.journey -> 0 ->> 'to' = 'redetermination'
+            AND NOT j.disk_evidence
+            AND j.claim_total::float >= 20000.00
+          SQL
+        end
+      end
+    end
+  end
+end
